requirements.txt
# python3.9
# pip install torch==2.2.2 torchvision==0.17.2 torchaudio==2.2.2 --index-url https://download.pytorch.org/whl/cu121
diffusers==0.31.0
transformers==4.51.3
xformers==0.0.25.post1 
accelerate==0.31.0
sentencepiece==0.2.0
tokenizers==0.21.4
bitsandbytes==0.47.0
detectron2==0.6
fvcore==0.1.5.post20221221
numba==0.60.0
numpy==1.26.3
omegaconf==2.3.0
opencv-python==4.6.0.66  
pytorch-lightning==2.2.0 
qwen-vl-utils==0.0.11

